refactor: refactor of code
