+++ src/components/FormRegistration/FormRegistration.tsx
@@ -2,83 +2,16 @@ import { useForm } from 'react-hook-form';
 import { Input } from '../ui/Input/Input';
 import styles from './FormRegistration.module.css';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { z, ZodType } from 'zod';
 import { useContext, useEffect, useState } from 'react';
 import { AccessTokenContext } from '../../context/AccessTokenContext';
 import { Loader } from '../../ui-kit/Loader/Loader';
-import { FormData } from '../../types/user/formData';
 import { Link, useNavigate } from 'react-router-dom';
 import { FormAddress } from '../FormAddress/FormAddress';
 import { CheckBox } from '../ui/CheckBox/CheckBox';
-import { createCustomer } from '../../services/createCustomer';
-import { loginCustomer } from '../../services/loginCustomer';
-import { addAddresses } from '../../services/addAddresses';
-
-const postalCodeRegex = {
-  US: /^\d{5}$/,
-  RU: /^\d{6}$/,
-};
-
-const addressSchema = {
-  streetName: z
-    .string()
-    .min(1, 'Street is required')
-    .regex(/^[A-Za-z0-9 ]+$/, 'Use only english letters'),
-  city: z
-    .string()
-    .min(1, 'City is required')
-    .regex(/^[A-Za-z\s'-]+$/, 'City must not contain numbers or special characters'),
-  postalCode: z
-    .string()
-    .refine((val) => postalCodeRegex.US.test(val) || postalCodeRegex.RU.test(val), {
-      message: 'Invalid postal code format for US or Russia',
-    }),
-
-  country: z.string(),
-};
-
-const formSchema: ZodType<FormData> = z.object({
-  email: z.string().email({ message: 'Invalid email address' }),
-  password: z
-    .string()
-    .min(8, 'Password must be at least 8 characters')
-    .regex(/[a-z]/, 'Must include a lowercase letter')
-    .regex(/[A-Z]/, 'Must include an uppercase letter')
-    .regex(/\d/, 'Must include a number'),
-  firstName: z
-    .string()
-    .min(1, 'First name is required')
-    .regex(/^[A-Za-z\s'-]+$/, 'First name must not contain numbers or special characters'),
-
-  lastName: z
-    .string()
-    .min(1, 'Last name is required')
-    .regex(/^[A-Za-z\s'-]+$/, 'Last name must not contain numbers or special characters'),
-
-  dateOfBirth: z.string().refine(
-    (val) => {
-      const dob = new Date(val);
-      if (isNaN(dob.getTime())) return false;
-      const now = new Date();
-      const age = now.getFullYear() - dob.getFullYear();
-      const hasBirthdayPassedThisYear =
-        now.getMonth() > dob.getMonth() ||
-        (now.getMonth() === dob.getMonth() && now.getDate() >= dob.getDate());
-      const adjustedAge = hasBirthdayPassedThisYear ? age : age - 1;
-      return adjustedAge >= 14;
-    },
-    {
-      message: 'You must be at least 14 years old',
-    },
-  ),
-  shippingAddress: z.object(addressSchema),
-  billingAddress: z.object(addressSchema),
-  isBillingSameAsShipping: z.boolean().optional(),
-  isBillingDefaultAddress: z.boolean().optional(),
-  isShippingDefaultAddress: z.boolean().optional(),
-  defaultShippingAddress: z.number().optional(),
-  defaultBillingAddress: z.number().optional(),
-});
+import { createCustomer } from '../../api/createCustomer';
+import { loginCustomer } from '../../api/loginCustomer';
+import { addAddresses } from '../../api/addAddresses';
+import { RegisterFormData, registerSchema } from '../../schemas/register.schema';
 
 export function FormRegistration() {
   const {
@@ -88,8 +21,8 @@ export function FormRegistration() {
     watch,
     setValue,
     formState: { errors, isValid },
-  } = useForm<FormData>({
-    resolver: zodResolver(formSchema),
+  } = useForm<RegisterFormData>({
+    resolver: zodResolver(registerSchema),
     mode: 'all',
     reValidateMode: 'onChange',
     defaultValues: {
@@ -133,7 +66,7 @@ export function FormRegistration() {
 
   const token = useContext(AccessTokenContext);
 
-  const onSumbit = async (data: FormData) => {
+  const onSumbit = async (data: RegisterFormData) => {
     const userData = isBillingSameAsShipping
       ? {
           ...data,
@@ -148,7 +81,7 @@ export function FormRegistration() {
       setIsLoading(true);
       const newCustomer = await createCustomer(userData, token);
       const tokenCustomer = await loginCustomer(userData);
-      await addAddresses(userData, newCustomer, tokenCustomer);
+      await addAddresses({ data: userData, customer: newCustomer, userToken: tokenCustomer });
       setIsLoading(false);
       reset();
       navigate('/');
